Fix build issue with fancy dependency. Move generate_delta and mask_payload_crt to fancy utils
