+++ popsicle/deltas.txt
@@ -1,1 +1,1 @@
-{"29":{"ModN":{"q":29,"ds":[1,10,1,4,2,28,22,11,5,2,17,22,26,0,8,8,28,17,13,11,8,19,24,21,8]}},"53":{"ModN":{"q":53,"ds":[1,43,44,42,9,24,42,5,31,8,23,11,41,21,12,15,39,41,31,23,7]}},"2":{"Mod2":{"val":{"block":323163145503925358191888121545623652407}}},"17":{"ModN":{"q":17,"ds":[1,6,12,10,4,2,3,13,6,3,11,4,11,0,15,16,10,6,7,12,2,9,6,14,8]}},"37":{"ModN":{"q":37,"ds":[1,33,33,28,8,2,32,14,8,33,30,32,4,34,15,3,15,1,18,27,11]}},"7":{"ModN":{"q":7,"ds":[1,5,3,6,2,6,0,6,4,5,0,1,5,6,3,1,0,1,2,6,0,2,6,1,6,4,5,5,5,1,0,2,2,3,2,2,0,4,6,2,5,4]}},"31":{"ModN":{"q":31,"ds":[1,29,6,8,15,28,8,8,23,22,28,4,27,30,6,28,3,26,30,22,14,19,14,8,1]}},"23":{"ModN":{"q":23,"ds":[1,8,5,8,16,16,22,10,2,10,10,6,3,3,2,7,7,9,12,17,10,1,3,10,12]}},"19":{"ModN":{"q":19,"ds":[1,10,4,1,4,4,4,7,17,14,8,13,11,10,6,3,8,6,5,10,11,17,3,16,16]}},"41":{"ModN":{"q":41,"ds":[1,1,27,1,10,3,36,28,12,40,40,9,38,27,20,2,23,23,25,0,40]}},"43":{"ModN":{"q":43,"ds":[1,36,0,10,33,9,8,28,34,34,35,16,33,41,37,6,32,5,37,24,32]}},"3":{"Mod3":{"lsb":4035792787901041033,"msb":9517709476397003824}},"11":{"ModN":{"q":11,"ds":[1,4,2,8,3,2,7,9,4,9,3,10,1,9,6,0,5,8,7,8,6,1,8,9,8,9,1,8,7,6,9,0]}},"13":{"ModN":{"q":13,"ds":[1,1,8,7,2,3,8,11,9,8,9,8,0,9,0,4,2,5,2,0,4,1,11,5,1,0,9,8,3,6,5,8]}},"5":{"ModN":{"q":5,"ds":[1,1,3,1,3,1,2,1,1,4,3,3,4,0,2,4,4,3,1,2,4,1,2,0,0,4,3,1,2,4,2,0,4,4,0,4,2,3,3,0,2,0]}},"47":{"ModN":{"q":47,"ds":[1,19,2,44,34,15,14,26,29,5,27,30,1,38,0,43,41,0,37,10,32]}}}
+{"3":{"Mod3":{"lsb":3214230044099060307,"msb":1396118946025183364}},"41":{"ModN":{"q":41,"ds":[1,10,14,33,1,9,13,20,10,15,7,9,0,36,24,5,33,18,29,19,5]}},"29":{"ModN":{"q":29,"ds":[1,16,5,6,11,0,19,9,28,28,26,6,6,3,0,4,24,12,18,23,2,18,22,8,12]}},"11":{"ModN":{"q":11,"ds":[1,9,7,4,9,6,6,10,0,10,1,7,1,2,3,7,2,9,3,8,5,0,6,8,4,0,3,2,10,1,1,1]}},"53":{"ModN":{"q":53,"ds":[1,7,36,47,32,14,12,31,22,40,44,19,48,49,0,52,50,46,27,25,36]}},"43":{"ModN":{"q":43,"ds":[1,9,21,10,13,13,13,39,18,39,8,2,14,21,10,14,22,36,20,22,1]}},"31":{"ModN":{"q":31,"ds":[1,24,1,26,0,4,19,9,0,26,19,4,14,12,0,7,10,19,16,5,6,6,1,10,1]}},"23":{"ModN":{"q":23,"ds":[1,5,14,3,19,5,1,5,6,4,3,19,12,6,17,5,11,8,7,8,5,13,14,1,11]}},"13":{"ModN":{"q":13,"ds":[1,0,2,7,5,10,6,8,12,3,7,0,4,1,9,9,7,0,3,9,7,4,5,1,5,4,4,9,1,5,8,10]}},"17":{"ModN":{"q":17,"ds":[1,13,3,15,8,10,6,0,9,8,14,4,15,4,1,8,16,4,4,10,16,15,13,13,14]}},"2":{"Mod2":{"val":{"block":192227484052000602439299336570232144927}}},"37":{"ModN":{"q":37,"ds":[1,1,31,23,4,1,25,33,18,5,20,12,1,29,1,22,21,18,9,5,9]}},"5":{"ModN":{"q":5,"ds":[1,1,2,3,3,2,1,3,4,3,3,1,4,0,2,2,2,4,0,3,3,1,1,0,0,3,0,0,2,0,3,1,3,0,4,2,3,3,3,0,2,0]}},"47":{"ModN":{"q":47,"ds":[1,14,14,37,9,37,9,24,9,40,8,30,5,0,44,3,21,15,3,39,29]}},"7":{"ModN":{"q":7,"ds":[1,5,1,2,1,0,0,2,3,0,1,4,5,5,6,3,5,5,5,0,1,5,0,3,1,2,1,4,3,4,0,6,6,0,2,2,4,5,3,6,6,3]}},"19":{"ModN":{"q":19,"ds":[1,8,14,14,8,7,9,10,2,7,9,14,11,3,18,17,5,0,15,16,16,7,11,8,1]}}}
