+++ popsicle/examples/match-compute-parallel/server/delta.txt
@@ -1,1 +1,1 @@
-{"13":{"ModN":{"q":13,"ds":[1,12,2,9,9,11,12,9,4,1,7,3,4,10,12,11,10,5,6,1,3,3,3,12,3,5,5,12,0,12,8,2]}},"31":{"ModN":{"q":31,"ds":[1,25,13,20,21,9,12,2,28,23,10,23,21,13,24,2,2,24,28,26,6,10,19,24,16]}},"41":{"ModN":{"q":41,"ds":[1,5,13,17,12,25,14,38,1,38,14,23,10,30,14,17,6,20,26,10,21]}},"17":{"ModN":{"q":17,"ds":[1,5,11,9,4,6,7,2,4,1,5,7,6,12,5,3,13,10,3,13,9,5,10,8,1]}},"23":{"ModN":{"q":23,"ds":[1,14,1,3,12,2,10,6,8,7,19,16,6,15,14,2,13,13,13,8,6,13,8,16,8]}},"43":{"ModN":{"q":43,"ds":[1,7,0,3,26,23,33,27,26,34,32,38,31,0,6,24,19,4,4,4,12]}},"5":{"ModN":{"q":5,"ds":[1,1,1,0,4,4,4,4,4,2,0,4,2,0,4,1,1,4,1,3,1,4,2,0,4,2,2,0,0,3,2,0,2,0,3,4,4,4,1,3,1,0]}},"7":{"ModN":{"q":7,"ds":[1,4,0,3,2,1,1,1,6,4,2,2,4,2,3,1,2,5,3,0,0,1,3,2,1,1,5,2,3,4,2,6,6,3,2,5,0,4,4,1,6,5]}},"2":{"Mod2":{"val":{"block":304045765093466680948374422778613109375}}},"3":{"Mod3":{"lsb":8116776807989203835,"msb":10163323007079558148}},"11":{"ModN":{"q":11,"ds":[1,5,8,7,7,5,6,10,10,5,3,7,6,5,9,8,0,10,7,5,7,2,3,5,5,6,0,7,2,4,9,7]}},"19":{"ModN":{"q":19,"ds":[1,8,14,8,4,16,16,1,17,0,9,4,13,8,3,13,12,17,12,6,15,13,2,2,7]}},"29":{"ModN":{"q":29,"ds":[1,2,23,7,19,21,18,6,12,3,12,19,16,26,25,18,5,19,13,27,7,25,9,8,19]}},"53":{"ModN":{"q":53,"ds":[1,28,9,48,44,9,26,7,50,25,4,37,18,34,39,19,7,34,23,25,43]}},"47":{"ModN":{"q":47,"ds":[1,45,15,38,10,24,45,6,27,26,33,3,1,33,45,8,17,29,22,41,6]}},"37":{"ModN":{"q":37,"ds":[1,28,0,32,4,28,35,22,27,16,13,8,8,2,2,6,13,21,18,9,20]}}}
+{"37":{"ModN":{"q":37,"ds":[1,35,23,34,1,3,28,12,29,12,13,3,10,30,22,11,22,32,34,34,28]}},"53":{"ModN":{"q":53,"ds":[1,39,26,47,3,14,25,19,49,32,51,6,29,44,39,16,6,14,3,11,10]}},"7":{"ModN":{"q":7,"ds":[1,6,6,0,6,6,2,1,5,5,4,4,0,5,2,4,1,2,2,0,2,4,6,2,6,6,1,3,3,3,2,6,2,0,4,4,6,2,2,5,1,3]}},"13":{"ModN":{"q":13,"ds":[1,2,10,8,5,0,11,3,10,7,2,9,6,2,4,2,9,2,9,11,2,6,0,0,11,8,4,0,12,2,5,9]}},"3":{"Mod3":{"lsb":9683182456686587033,"msb":8760170134187983428}},"29":{"ModN":{"q":29,"ds":[1,4,23,8,17,28,1,1,5,28,24,15,21,12,18,24,19,19,28,12,1,8,9,28,22]}},"11":{"ModN":{"q":11,"ds":[1,4,3,4,9,0,9,8,3,3,1,1,1,3,8,4,3,2,3,5,7,6,7,7,1,8,5,9,3,0,7,5]}},"23":{"ModN":{"q":23,"ds":[1,12,5,12,7,13,16,17,19,15,20,21,4,18,16,17,10,14,8,17,1,6,1,11,1]}},"41":{"ModN":{"q":41,"ds":[1,29,26,14,27,0,34,37,25,5,24,7,25,12,30,3,4,27,23,10,10]}},"47":{"ModN":{"q":47,"ds":[1,30,8,36,35,40,9,39,0,44,31,34,31,34,39,31,17,33,39,1,40]}},"17":{"ModN":{"q":17,"ds":[1,8,14,10,5,13,1,7,2,15,15,6,12,1,6,1,8,15,16,0,3,0,4,9,9]}},"2":{"Mod2":{"val":{"block":90380868176781357686186420101804008719}}},"43":{"ModN":{"q":43,"ds":[1,0,4,36,32,2,16,4,3,38,6,33,30,11,34,9,2,3,16,17,41]}},"19":{"ModN":{"q":19,"ds":[1,15,16,11,10,1,1,11,0,11,2,11,2,5,11,2,1,14,15,4,6,0,10,2,9]}},"5":{"ModN":{"q":5,"ds":[1,3,3,4,4,2,1,2,1,4,0,2,0,4,2,4,4,1,2,0,3,3,2,0,0,4,2,2,3,1,4,0,4,3,1,0,1,4,2,2,1,2]}},"31":{"ModN":{"q":31,"ds":[1,7,18,23,9,8,26,25,6,24,16,13,26,23,16,9,24,2,18,6,6,1,10,12,2]}}}
