+++ popsicle/src/utils.rs
@@ -5,10 +5,8 @@
 // See LICENSE for licensing information.
 
 use rand::{CryptoRng, Rng};
-use scuttlebutt::{AesHash, Block, Block512};
+use scuttlebutt::{AesHash, Block};
 use sha2::{Digest, Sha256};
-use fancy_garbling::Wire;
-use std::collections::HashMap;
 
 // Compress an arbitrary vector into a 128-bit chunk, leaving the final 8-bits
 // as zero. We need to leave 8 bits free in order to add in the hash index when
@@ -46,56 +44,6 @@ pub fn rand_vec<RNG: CryptoRng + Rng>(n: usize, rng: &mut RNG) -> Vec<u8> {
 pub fn rand_vec_vec<RNG: CryptoRng + Rng>(n: usize, m: usize, rng: &mut RNG) -> Vec<Vec<u8>> {
     (0..n).map(|_| rand_vec(m, rng)).collect()
 }
-
-pub fn generate_deltas(primes: &[u16]) -> HashMap<u16, Wire> {
-    let mut deltas = HashMap::new();
-    let mut rng = rand::thread_rng();
-    for q in primes{
-        deltas.insert(*q, Wire::rand_delta(&mut rng, *q));
-    }
-    deltas
-}
-
-
-//Assumes payloads are up to 64bit long
-pub fn mask_payload_crt<RNG: Rng + CryptoRng>(x: Block512, y: Block512, rng:&mut RNG)
-        -> Block512{
-
-    let x_crt = block512_to_crt(x);
-    let y_crt = block512_to_crt(y);
-    let q = fancy_garbling::util::primes_with_width(64);
-    let mut res_crt = Vec::new();
-    for i in 0..q.len(){
-        res_crt.push((x_crt[i]+y_crt[i]) % q[i]);
-    }
-    let res = fancy_garbling::util::crt_inv(&res_crt, &q).to_le_bytes();
-    let mut block = [0 as u8; 64];
-    for i in 0..64{
-        if i < res.len(){
-            block[i] = res[i];
-        }else{
-            block[i] = rng.gen::<u8>();
-        }
-    }
-    Block512::from(block)
-}
-
-//Assumes payloads are up to 64bit long i.e 8 bytes
-fn block512_to_crt(b: Block512) -> Vec<u16>{
-    let b_val = b.prefix(8);
-
-    let mut b_128 = [0 as u8; 16];
-
-    // Loop over the 8 bytes of 64b b_val
-    for i in 0..8{
-        b_128[i] = b_val[i];
-    }
-
-    let q = fancy_garbling::util::primes_with_width(64);
-    let b_crt = fancy_garbling::util::crt(u128::from_le_bytes(b_128), &q);
-    b_crt
-}
-
 
 #[cfg(test)]
 mod tests {
