--- popsicle/values.txt
@@ -1,4 +1,0 @@
-5
-6,8,9
-7,1,10
-8,4,11
